fix: upgrade contract interface, emit all params on init for off-chain verification
initiate event now emits all swap params, which need to be verified
off-chain for security.
similarly, refund and instant refund events emit the secret hash
instead of swap id (it has been totally decoupled from the contract).
this secret hash can be used as the unique identifier for the swap.
even the PDA is now derived from the secret hash.