Standardized event structure, doc comments and minor refactors (#4)
* refactor: Rename `Initiated` event structure to match with SPL

This is to have a common events structure for both native program and
spl program.

* refactor: Replace literal with const `ANCHOR_DISCRIMINATOR`

* refactor: Use anchor provided `InitSpace` for specifying PDA size

Anchor now provides an attribute macro `InitSpace` that adds a
constant `INIT_SPACE` which specifies the size that needs to be
allocated for a PDA during creation.
Thus, use this instead of `std::mem::size_of`

* fix: Initialize the on-chain swap data after the transfer

The transfer instruction is a fallible operation which can fail
e.g when the user doesn't have enough balance.
Initializing the swap data after this operation would save
compute during such cases.

* refactor: Minor simplification to error strings

* doc: Add documentation comments