feat: Introduce Initiate on behalf
Initiates typically involve an initiator depositing funds, thus
requiring their signature.
For cases where an initiator wants to participate but cannot provide
their signature, they can do so using initiate on behalf, where a
party referred to as the "funder" deposits funds on the initiator's
behalf. Except for this, the role of the initiator stays the same with
regards to refunds.